fix: only admins can change community name and places can be removed only if you own it

+++ src/logic/community/places.ts
@@ -133,9 +133,8 @@ export async function createCommunityPlacesComponent(
 
     removePlace: async (communityId: string, userAddress: EthAddress, placeId: string): Promise<void> => {
       await validateCommunityExists(communityId)
-
       await validatePlaceExists(communityId, placeId)
-
+      await validateOwnership([placeId], userAddress)
       await communityRoles.validatePermissionToRemovePlacesFromCommunity(communityId, userAddress)
 
       await communitiesDb.removeCommunityPlace(communityId, placeId)
